refactor to use single form context
